--- src/pymmcore_gui/_writers/__init__.py
@@ -1,5 +1,0 @@
-"""Writer classes for different file formats."""
-
-from ._tensorstore_zarr import _TensorStoreHandler
-
-__all__ = ["_TensorStoreHandler"]
