+++ src/pymmcore_gui/io/__init__.py
@@ -1,0 +1,4 @@
+from ._writers._tensorstore_zarr import TensorStoreHandler
+from .readers import OMEZarrReader, TensorstoreZarrReader
+
+__all__ = ["OMEZarrReader", "TensorStoreHandler", "TensorstoreZarrReader"]
