+++ src/pymmcore_gui/io/readers/__init__.py
