+++ src/pymmcore_gui/io/readers/_ome_zarr_reader.py
