+++ src/pymmcore_gui/io/readers/_tensorstore_zarr_reader.py
