+++ src/pymmcore_gui/widgets/_install_widget.py
