+++ src/pymmcore_gui/widgets/_mm_console.py
