+++ src/pymmcore_gui/widgets/_shutters_toolbar.py
