+++ src/pymmcore_gui/widgets/_stage_control.py
